:test_tube: Refactors Brand case to Marketplaces case
