+++ apps/ex_commerce_web/test/support/resource_cases/marketplaces_case.ex
@@ -1,4 +1,4 @@
-defmodule ExCommerceWeb.ResourceCases.BrandCase do
+defmodule ExCommerceWeb.ResourceCases.MarketplacesCase do
   @moduledoc """
   This module defines the test case to be used by
   tests that require brands helpers.
@@ -13,6 +13,7 @@ defmodule ExCommerceWeb.ResourceCases.BrandCase do
       alias ExCommerce.Accounts.User
       alias ExCommerce.Marketplaces
       alias ExCommerce.Marketplaces.Brand
+      alias ExCommerce.Marketplaces.Shop
 
       defp assoc_user_brand(%{
              user: %User{id: user_id},
@@ -26,6 +27,16 @@ defmodule ExCommerceWeb.ResourceCases.BrandCase do
 
         %{}
       end
+
+      defp assoc_brand_shop(%{brand: %Brand{id: brand_id}, shop: %Shop{} = shop}) do
+        {:ok, shop} =
+          Marketplaces.update_shop(shop, %{
+            brand_id: brand_id
+          })
+
+        %{shop: shop}
+      end
+
     end
   end
 end
